requirements.txt
streamlit
openai
pypdf
requests
huggingface-hub
numpy
scikit-learn
python-dotenv

